Add initial CMake support
Only covers imgui_frontend up to now, not teensy_backend, and the result of
the build is untested.
It also needs some comments to explain the libraries that are linked, and to
note that some of these are platform dependant so will need handling
differently to get this to work on other platforms.

build/ is added to the list of things git should ignore.