Eliminate redundant comma operators
The return values from various ImGui input functions were being assigned 
to variables on the right of a comma operator; to the left, and thus 
ignored, was the variable for the actual output of the function which in 
C/C++ is actually passed by reference using a pointer argument. Maybe 
because whoever wrote it was looking at Python examples where a tuple is 
returned?