Make the backend constants.h & constants_alive.h be the same
None of the constants_targetname.h files are actually included, only 
constants.h (and constants_global.h). The intention was, that to build a 
target-specific elf file, the appropriate constants_targetname.h file, 
e.g. constants_alive.h, would first be copied over the constants.h file. 
As stored in the repository, the constants.h file was almost, but not 
entirely the same as constants_alive.h. This commit makes them the same 
before we go on in a subsequent commit to change constants.h to pick the 
appropriate constants_targetname.h file according to the 
TARGET_${TARGET} macros. I've put this commit in as a separate step to 
make it easier to see that the changes made to constants_alive.h and 
constants.h to get them back to a matching state are correct.