Fix 'declared but not defined' warning in ahpc build
aerInputEncoder_ISR() was being forward declared not protected by #ifdef 
EXISTS_ENCODER, even though the definition was so protected. AHPC does 
not #define EXISTS_ENCODER, whereas other targets do, thus in that case 
giving a warning. This commit adds #ifdef protection around the forward 
declaration.

It might be better to rearrange stuff to avoid the forward declaration, 
but I don't want to go down that route now in case it turns out to be 
even a small rabbit hole.