Remove redundant (and wrong) definition of BIASGEN_PACKET_SIZE
If we need to know the size of the biasgen packet (or anything else) we 
should get it by using sizeof() on it such that the value can't get out 
of sync with the structure, as appears to have happened in this case. 
Fortunately it was not used anywhere except in a commented out bit of 
debugging code, which is also removed.