Add support in backend for setting Texel biases
In response to a Pkt_setBiasGen packet, call the new Texel::setBias() 
function which just ORs in TEXEL_AE_REGWRITE and puts address and value 
in the right position in the word and writes the resulting value to the 
Texel chip. This obviously relies on the frontend providing reasonable 
values in the BIASGEN_command biasNo and currentValue_binary fields, 
which will depend on amongst other things a correct implementation of 
getBiasGenPacket() in dataFunctions.cpp for TARGET_TEXEL. This is still 
a TODO item.